Add a world container object
This is important, and points the way at future functionality. Before we only had Genesis create a single WorldRoot layer and populate it with tiles. If you changed location on the map, it'd destroy the world. In doing this, we now create the ability for the GenesisController to build multiple worlds with their own data layers simultaneously...